Passport, API and AJAX

+++ controllers/videoController.js
@@ -1,5 +1,6 @@
 import routes from "../routes";
 import Video from "../models/Video";
+import Comment from "../models/Comment";
 
 // export const home = (req, res) => {
 //     res.render("home", {pageTitle: "Home", videos});
@@ -40,9 +41,8 @@ export const videoDetail = async (req, res) => {
     } = req;
 
     try {
-        // .findById() 는 mongoose의 Querys 중 하나
-        const video = await Video.findById(id);
-        console.log(video);
+        const video = await (await Video.findById(id).populate("creator").populated("comment"));
+        // console.log(video);
         res.render("videoDetail", { pageTitle: video.title, video });
     } catch (error) {
         console.log(error);
@@ -67,7 +67,8 @@ export const postUpload = async (req, res) => {
     const newVideo = await Video.create({
         fileUrl: path,
         title,
-        description
+        description,
+        creator: req.user.id
     });
 
     console.log(newVideo);
@@ -83,7 +84,11 @@ export const getEditVideo = async (req, res) => {
 
     try {
         const video = await Video.findById(id);
-        res.render("editVideo", {pageTitle: `Edit ${video.title}`, video});
+        if(String(video.creator) !== req.user.id) {
+            throw Error();
+        } else {
+            res.render("editVideo", {pageTitle: `Edit ${video.title}`, video});
+        }
     } catch (error) {
         res.redirect(routes.home);
     }
@@ -99,9 +104,15 @@ export const postEditVideo = async (req, res) => {
     } = req;
 
     try {
-        await Video.findOneAndUpdate({_id:id}, {title,description});
+        const video = await Video.findById(id);
+        if(String(video.creator) !== req.user.id) {
+            throw Error();
+        } else {
+            await Video.findOneAndUpdate({_id:id}, {title,description});
+        }
         res.redirect(routes.videoDetail(id));
     } catch (error) {
+        console.log(error);
         res.redirect(routes.home);
     }
 };
@@ -112,8 +123,53 @@ export const deleteVideo = async (req, res) => {
     } = req;
 
     try {
-        await Video.findOneAndRemove({_id:id});
+        const video = await Video.findById(id);
+        if(video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({_id:id});
+        }
     } catch (error) {
         res.redirect(routes.home);
+    }
+};
+
+export const postRegisterView = async (req, res) => {
+    const {
+        params: { id }
+    } = req;
+    try {
+        const video = await Video.findById(id);
+        // models/Video.js의 video의 views
+        video.views += 1;
+        video.save();
+        res.status(200);
+    } catch (error) {
+        res.status(400);
+    } finally {
+        res.end();
+    }
+};
+
+//Add Comment
+export const postAddComment = async (req, res) => {
+    const {
+        params: { id },
+        body: { comment },
+        user
+    } = req;
+    try {
+        const video = await Video.findById(id);
+        const newComment = await Comment.create({
+            text: comment,
+            creator: user.id
+        });
+        video.comments.push(newComment._id);
+        video.save();
+        res.send({commentData:newComment});
+    } catch (error) {
+        res.status(400);
+    } finally {
+        res.end();
     }
 };
